+++ src/oci/opsi/models/host_resource_statistics.py
@@ -120,7 +120,7 @@ class HostResourceStatistics(object):
     def capacity(self):
         """
         **[Required]** Gets the capacity of this HostResourceStatistics.
-        The maximum allocated amount of the resource metric type  (CPU, STORAGE).
+        The maximum allocated amount of the resource metric type  (CPU, STORAGE) for a set of databases.
 
 
         :return: The capacity of this HostResourceStatistics.
@@ -132,7 +132,7 @@ class HostResourceStatistics(object):
     def capacity(self, capacity):
         """
         Sets the capacity of this HostResourceStatistics.
-        The maximum allocated amount of the resource metric type  (CPU, STORAGE).
+        The maximum allocated amount of the resource metric type  (CPU, STORAGE) for a set of databases.
 
 
         :param capacity: The capacity of this HostResourceStatistics.