+++ src/oci/opsi/models/macs_managed_cloud_host_insight_summary.py
@@ -247,6 +247,7 @@ class MacsManagedCloudHostInsightSummary(HostInsightSummary):
         Gets the platform_type of this MacsManagedCloudHostInsightSummary.
         Platform type.
         Supported platformType(s) for MACS-managed external host insight: [LINUX].
+        Supported platformType(s) for MACS-managed cloud host insight: [LINUX].
         Supported platformType(s) for EM-managed external host insight: [LINUX, SOLARIS, SUNOS, ZLINUX].
 
         Allowed values for this property are: "LINUX", "SOLARIS", "SUNOS", "ZLINUX", 'UNKNOWN_ENUM_VALUE'.
@@ -264,6 +265,7 @@ class MacsManagedCloudHostInsightSummary(HostInsightSummary):
         Sets the platform_type of this MacsManagedCloudHostInsightSummary.
         Platform type.
         Supported platformType(s) for MACS-managed external host insight: [LINUX].
+        Supported platformType(s) for MACS-managed cloud host insight: [LINUX].
         Supported platformType(s) for EM-managed external host insight: [LINUX, SOLARIS, SUNOS, ZLINUX].