+++ src/oci/opsi/models/tablespace_usage_trend.py
@@ -99,7 +99,7 @@ class TablespaceUsageTrend(object):
     def capacity(self):
         """
         **[Required]** Gets the capacity of this TablespaceUsageTrend.
-        The maximum allocated amount of the resource metric type  (CPU, STORAGE).
+        The maximum allocated amount of the resource metric type  (CPU, STORAGE) for a set of databases.
 
 
         :return: The capacity of this TablespaceUsageTrend.
@@ -111,7 +111,7 @@ class TablespaceUsageTrend(object):
     def capacity(self, capacity):
         """
         Sets the capacity of this TablespaceUsageTrend.
-        The maximum allocated amount of the resource metric type  (CPU, STORAGE).
+        The maximum allocated amount of the resource metric type  (CPU, STORAGE) for a set of databases.
 
 
         :param capacity: The capacity of this TablespaceUsageTrend.